pyproject.toml
[tool.poetry]
name = "todo_list"
version = "0.1.0"
description = ""
authors = ["Анна Полякова <Sova040888@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.7"
split-settings = "^1.0.0"
drf-spectacular = "^0.26.5"
django-rest-framework = "^0.1.0"
drf-spectacular-sidecar = "^2023.10.1"
load-dotenv = "^0.1.0"
isort = "^5.12.0"
black = "^23.11.0"
django-environ = "^0.11.2"
pytest = "^7.4.3"
pytest-django = "^4.7.0"
djangorestframework-simplejwt = "^5.3.0"
psycopg2 = "^2.9.9"
gunicorn = "^21.2.0"
flake8 = "^6.1.0"
django-filter = "^23.4"
factory-boy = "^3.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 79
